codeweaver/agent.py
"""
Autonomous coding agent implementation
"""
import os
from typing import List, Optional, Dict, Any
from dataclasses import dataclass
import pytest
from dataclasses import dataclass

@dataclass
class CodingTask:
    description: str
    requirements: List[str]
    test_requirements: Optional[List[str]] = None

class CodingAgent:
    """An autonomous coding agent that can write and test code"""
    
    def __init__(self):
        self.api_key = os.getenv("DEEPSEEK_API_KEY")
        if not self.api_key:
            raise ValueError("DEEPSEEK_API_KEY environment variable not set")
        
        # Will initialize model connection in future PR
        self.model = None
        self.system_prompt = (
            "You are an autonomous coding agent. Your tasks are to:"
            "\n- Write code based on requirements"
            "\n- Create pytest tests"
            "\n- Evaluate code quality"
            "\n- Suggest improvements"
        )
        self.messages = [("system", self.system_prompt)]

    def add_message(self, message: str, role: str = "human") -> None:
        """Add a message to the conversation history"""
        if role not in ["system", "human", "ai"]:
            raise ValueError(f"Invalid role: {role}")
        self.messages.append((role, message))

    def create_code(self, task: CodingTask) -> str:
        """Generate code based on task requirements"""
        # Will implement in future PR
        return ""

    def create_tests(self, code: str, task: CodingTask) -> str:
        """Generate pytest tests for the code"""
        # Will implement in future PR
        return ""

    def evaluate_code(self, code: str, tests: str) -> Dict[str, Any]:
        """Evaluate code quality and test results"""
        # Will implement in future PR
        return {}

    def improve_code(self, code: str, evaluation: Dict[str, Any]) -> str:
        """Suggest and implement code improvements"""
        # Will implement in future PR
        return ""
"""
Autonomous coding agent implementation
"""
from dataclasses import dataclass
import os

@dataclass
class CodingTask:
    """A coding task to be performed by the agent"""
    description: str
    language: str

class CodingAgent:
    """An autonomous coding agent"""
    
    def __init__(self):
        """Initialize the coding agent"""
        self.api_key = os.getenv("DEEPSEQ_API_KEY")
        
    async def generate(self, task: CodingTask) -> str:
        """Generate code for the given task"""
        # TODO: Implement actual code generation
        return "def add(a, b):\n    return a + b"
